Disable smoothing on all raster sources (#485)
* Disable smoothing on all raster sources

Right now we don't know in advance whether someone is going to visualize
raster observation data, for which preserving grid cell boundaries is
important, or stylistic raster data like a hillshade, or rasterized
vector data. This change prioritizes the first use case at the expense
of the others, where smoothing may be useful, but in my view not
critical.

* Lint

Yay trailing commas, boo this linting rule :)

* Remove unused argument

* Lint