Avoid flashing an error screen when loading symbology menu
Because `featuresLoading` has an initial value of `{}`, subcomponents
would initially flash an error message, which is kinda gross.

Now, we check `isLoading` first. This had a strange side effect of
unstyling the RenderType dropdown! Somehow the `<select>` was being
wrapped in a div automatically, but now that's not happening. Manually
added the wrapper div, following patterns from elsewhere in the
codebase. Also added `jp-mod-styled` class, again following patterns
from elsewhere in the codebase.