Access props from a props argument
I feel that while this is more verbose, it reduces cognitive load when
reading the component code. Previously, in order to know where a value
comes from (i.e. props vs state), I needed to remember the names of all
the props. With this change, I don't need to know that because I can see
the value is accessed from the props object.