Add pre-processing script to enable our code generation tools to work with $refs
The root problem is that our TS-type-generation tool,
jsonschema-to-typescript, expects $ref paths to be relative to the
schema root directory, while our Python-type-generation tool,
datamodel-codegen, expects $ref paths to be relative to the file
containing the $ref.

To work around this, I'm adding a pre-processing step to the Python type
generation code which converts the paths to look like datamodel-codegen
expects.