Restore follow mode (#812)
* Attempt to fix follow

The idea is that we'd pass in a custom icon renderer function to the
UsersItem component. Our custom function is based on the old code prior
to [PR #723](https://github.com/geojupyter/jupytergis/pull/723).

I really don't know what I'm doing here. ToolbarWidget isn't a react
component, so I don't know how to manage state! Perhaps we don't need to
keep track of state at all, we just need to update the model and let the
model keep track of it.

Co-authored-by: Arjun Verma <arjxn.py@gmail.com>

* bring back follow mode

* pointer and border for user item

---------

Co-authored-by: Arjun Verma <arjxn.py@gmail.com>